refactor : enum 변경

+++ src/main/java/com/example/toanyone/domain/store/repository/StoreRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Long> {
     Optional<Store> findById(Long storeId);
 
     default Store findByIdOrElseThrow(Long storeId) {
